requirements.txt
streamlit
ultralytics
tensorflow
deepface
opencv-python
numpy
Pillow